requirements.txt
Pillow==10.2.0
pyautogui==0.9.54
schedule==1.2.1
python-dotenv==1.0.1
openai==1.12.0
requests==2.31.0
cloudinary==1.41.0 